move: UserRole to user domain
