+++ src/main/java/com/dku/council/global/auth/jwt/JwtProvider.java
@@ -1,6 +1,6 @@
 package com.dku.council.global.auth.jwt;
 
-import com.dku.council.domain.UserRole;
+import com.dku.council.domain.user.UserRole;
 import com.dku.council.domain.user.User;
 import com.dku.council.global.error.exception.ExpiredTokenException;
 import com.dku.council.global.error.exception.IllegalTypeException;
